scripts/requirements.txt
deepl
python-dotenv
